+++ tests/__init__.py
@@ -3,7 +3,7 @@
 import os
 
 
-def load_fixture(filename):
+def load_fixture(filename: str):
     """Load a fixture."""
     path = os.path.join(os.path.dirname(__file__), "fixtures", filename)
     with open(path, encoding="utf-8") as fp: